Add tests for all CSV versions
Tests that depend on a specific version of the requirements are split
into separate files, since the main test file was growing too large.

For pre-v2.2.0 versions, there are no modifier rows. Therefore, code
information is always required.
